rename renderText.* to gui_renderText.*
