+++ CMakeLists.txt
@@ -68,7 +68,9 @@ add_executable(autobleem-gui
         src/code/gui/gui_manager.h
         src/code/gui/gui_editor.cpp
         src/code/gui/gui_editor.h
-        src/code/gui/gui_selectmemcard.cpp
+		src/code/gui/gui_renderText.h
+		src/code/gui/gui_renderText.cpp
+		src/code/gui/gui_selectmemcard.cpp
         src/code/gui/gui_selectmemcard.h
 		src/code/ver_migration.cpp
 		src/code/ver_migration.h
@@ -109,9 +111,7 @@ add_executable(autobleem-gui
         src/code/launcher/gui_btn_guide.cpp
 		src/code/launcher/gui_btn_guide.h
 		src/code/engine/padmapper.cpp
-		src/code/engine/padmapper.h
-		src/code/gui/renderText.h
-		src/code/gui/renderText.cpp)
+		src/code/engine/padmapper.h)
 
 add_executable(starter
 		src/code/starter.cpp
